requirements/dev.txt
-r test.txt
# Code quality
# ------------------------------------------------------------------------------
radon==6.0.1
pytype==2023.4.27
pre-commit==3.3.1
pydocstyle==6.3.0
mypy==1.3.0
django_stubs==4.2.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10

# Docs
# ------------------------------------------------------------------------------
mkdocs==1.4.3
mkdocstrings-python==1.1.0
mkdocstrings==0.22.0
mkdocs-material==9.1.15
mkdocs-gen-files==0.5.0
mkdocs-literate-nav==0.6.0
mkdocs-section-index==0.3.5
babel==2.12.1
